Work/record merge history

+++ src/libraries/Gaspra.SqlGenerator/Factories/ScriptVariableFactory.cs
@@ -28,40 +28,49 @@ namespace Gaspra.SqlGenerator.Factories
                 {
                     try
                     {
-                        var scriptFileName = $"{schema.Name}.Merge{table.Name}.sql";
-
-                        var scriptName = $"Merge{table.Name}";
-
-                        var tableTypeVariableName = $"{table.Name}";
-
-                        var tableTypeName = $"TT_{table.Name}";
-
-                        var tableTypeColumns = table.TableTypeColumns(schema);
-
-                        var mergeIdentifierColumns = table.MergeIdentifierColumns(schema);
-
-                        var deleteIdentifierColumns = table.DeleteIdentifierColumns(schema);
-
-                        var retentionPolicy = table.RetentionPolicy();
-
-                        var tablesToJoin = table.TablesToJoin(schema);
-
-                        var mergeScriptVariableSet = new MergeScriptVariableSet
+                        if (!table.ShouldSkipTable())
                         {
-                            Schema = schema,
-                            Table = table,
-                            DeleteIdentifierColumns = deleteIdentifierColumns,
-                            MergeIdentifierColumns = mergeIdentifierColumns,
-                            RetentionPolicy = retentionPolicy,
-                            ScriptFileName = scriptFileName,
-                            ScriptName = scriptName,
-                            TablesToJoin = tablesToJoin,
-                            TableTypeColumns = tableTypeColumns,
-                            TableTypeName = tableTypeName,
-                            TableTypeVariableName = tableTypeVariableName
-                        };
+                            var scriptFileName = $"{schema.Name}.Merge{table.Name}.sql";
 
-                        mergeScriptVariableSets.Add(mergeScriptVariableSet);
+                            var scriptName = $"Merge{table.Name}";
+
+                            var tableTypeVariableName = $"{table.Name}";
+
+                            var tableTypeName = $"TT_{table.Name}";
+
+                            var tableTypeColumns = table.TableTypeColumns(schema);
+
+                            var mergeIdentifierColumns = table.MergeIdentifierColumns(schema);
+
+                            var deleteIdentifierColumns = table.DeleteIdentifierColumns(schema);
+
+                            var retentionPolicy = table.RetentionPolicy();
+
+                            var tablesToJoin = table.TablesToJoin(schema);
+
+                            var mergeScriptVariableSet = new MergeScriptVariableSet
+                            {
+                                Schema = schema,
+                                Table = table,
+                                DeleteIdentifierColumns = deleteIdentifierColumns,
+                                MergeIdentifierColumns = mergeIdentifierColumns,
+                                RetentionPolicy = retentionPolicy,
+                                ScriptFileName = scriptFileName,
+                                ScriptName = scriptName,
+                                TablesToJoin = tablesToJoin,
+                                TableTypeColumns = tableTypeColumns,
+                                TableTypeName = tableTypeName,
+                                TableTypeVariableName = tableTypeVariableName
+                            };
+
+                            mergeScriptVariableSets.Add(mergeScriptVariableSet);
+                        }
+                        else
+                        {
+                            _logger.LogInformation("[{schema}].[{table}] has the extended property to skip, it won't be calculated in the merge variables",
+                                schema.Name,
+                                table.Name);
+                        }
                     }
                     catch (Exception ex)
                     {
